chore: removed v1 calendar view
