+++ apps/frontend/src/lib/calendar/CalendarIndex.svelte
@@ -3,10 +3,9 @@
 	import type { ICalendarField, IFieldType } from '@undb/core'
 	import type { ComponentType } from 'svelte'
 	import CalendarConfig from './CalendarConfig.svelte'
-	import DateCalendarView from './DateCalendarView.svelte'
 	import DateRangeCalendarView from './DateRangeCalendarView.svelte'
 	import * as Card from '$lib/components/ui/card'
-	import DateCalendarView2 from './DateCalendarView2.svelte'
+	import DateCalendarView from './DateCalendarView.svelte'
 
 	const table = getTable()
 	const view = getView()
@@ -15,7 +14,7 @@
 	$: field = fieldId ? ($table.schema.getFieldById(fieldId).into() as ICalendarField | undefined) : undefined
 
 	const map: Partial<Record<IFieldType, ComponentType>> = {
-		date: DateCalendarView2,
+		date: DateCalendarView,
 		'date-range': DateRangeCalendarView,
 	}
 </script>
